+++ actions/create.c
@@ -1,8 +1,8 @@
-#include "create.h"
-#include "argc.h"
-#include "format.h"
-#include "gible.h"
-#include "utils.h"
+#include "actions/create.h"
+#include "helpers/argc.h"
+#include "helpers/format.h"
+#include "helpers/strings.h"
+#include "helpers/utils.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -41,8 +41,9 @@ int gible_create(const char *execname, int argc, char *argv[])
     char *bfn = parser.positional[1];
     char *ofn = parser.positional[2];
 
-    if (are_filenames_same(pfn, bfn, ofn))
-        return 1;
+    int ret;
+    if ((ret = are_filenames_same(pfn, bfn, ofn)))
+        return (gible_error(same_filename_errors[ret - 1]), 1);
 
     if (!file_exists(pfn))
         return (gible_error("Patched file does not exist."), 1);
