+++ formats/ups.c
@@ -1,8 +1,8 @@
-#include "../bytearray.h"
-#include "../crc32.h"
-#include "../filemap.h"
-#include "../format.h"
-#include "../utils.h"
+#include "helpers/bytearray.h"
+#include "helpers/crc32.h"
+#include "helpers/filemap.h"
+#include "helpers/format.h"
+#include "helpers/utils.h"
 #include <stdint.h>
 #include <stdio.h>
 #include <string.h>
