--- gible.h
@@ -1,6 +1,0 @@
-#ifndef GIBLE_H
-#define GIBLE_H
-
-extern const char gible_description[];
-
-#endif // GIBLE_H
