+++ helpers/crc32.h
@@ -1,6 +1,6 @@
-#ifndef CRC32_H
-#define CRC32_H
+#ifndef HELPERS_CRC32_H
+#define HELPERS_CRC32_H
 
 unsigned int crc32(const void *data, unsigned long length, unsigned int prev);
 
-#endif /* CRC32_H */
+#endif /* HELPERS_CRC32_H */
