+++ helpers/strings.c
@@ -1,0 +1,9 @@
+#include "helpers/strings.h"
+
+const char gible_description[] = "Yet another rom patcher.";
+const char *same_filename_errors[] = 
+{
+    "Patch and input filenames are the same.",
+    "Input and output filenames are the same.",
+    "Patch and output filenames are the same."
+};
