+++ helpers/strings.h
@@ -1,0 +1,7 @@
+#ifndef HELPERS_STRINGS_H
+#define HELPERS_STRINGS_H
+
+extern const char gible_description[];
+extern const char *same_filename_errors[];
+
+#endif // HELPERS_STRINGS_H
