+++ helpers/utils.h
@@ -1,5 +1,5 @@
-#ifndef UTIL_H
-#define UTIL_H
+#ifndef HELPERS_UTIL_H
+#define HELPERS_UTIL_H
 
 #include <stdint.h>
 #include <stdio.h>
@@ -12,4 +12,4 @@ unsigned long readvint(unsigned char **stream);
 unsigned int read32le(const unsigned char *ptr);
 int are_filenames_same(const char *pfn, const char *ifn, const char *ofn);
 
-#endif /* UTIL_H */
+#endif /* HELPERS_UTIL_H */
